requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2025.1.31
charset-normalizer==3.4.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.1
cssutils==2.11.1
cycler==0.12.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
ftfy==6.3.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
iopath==0.1.10
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
ll-xist==5.78
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
portalocker==3.1.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycoco==0.7.2
pycocotools==2.0.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.30
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.110
ultralytics-thop==2.0.14
urllib3==2.4.0
wcwidth==0.2.13
